pcu_sock: Only announce dyn TS already configured by lower layers
Avoid announcing to PCU as available  a dyn TS not yet fully
configured in the phy. Otherwise when we receive the Chan Activation
over the PCU sock while pchan_is still is not PDCH, and we fail to fully
activate the channel at that time.
See previous commit description for more information on the issue.

We still want to check for pchan_want because we actually want to stop
announcing the TS when it is in progress of being changed to TCH. This
configuration change is continued/finished once we receive the resulting
Release from PCU.

Change-Id: I8e2b170c1f94e7dfe2576a1fc899bf9c8a826a44