sysmobts-mgr: Fix path to hwmon in /sys
The path to the temperature sensor was wrong which
results in the SysmoBTS manager always considering
that the state is critical and hence powering OFF
the PA on startup. (with the default config)

As a further side-effect, the temperatures were not
available in the vty 'show manager' command.

Redacted log previous to this patch:

  sysmobts_misc.c:128 Error reading temperature
  sysmobts_mgr_2050.c:323 Going to enable the PA.
  sysmobts_mgr_temp.c:252 Failed to read the digital temperature. rc=-1
  sysmobts_mgr_temp.c:267 Failed to read the RF temperature. rc=-1
  sysmobts_mgr_temp.c:203 Moving from state NORMAL to CRITICAL.
  sysmobts_mgr_temp.c:185 System has reached critical warning.
  sysmobts_mgr_temp.c:236 OML alert sent: 0

and in the vty:

  Temperature control state: CRITICAL
  Current Temperatures
   Digital: -0.001000 Celsius
   RF:      -0.001000 Celsius

With this patch we correctly enable the PA and now can show in the vty:

  Temperature control state: NORMAL
  Current Temperatures
   Digital: 39.000000 Celsius
   RF:      35.125000 Celsius

Change-Id: Iac8e9a866ad6326874af91bd12a981f39a9a24c7